Workaround to calculate root folder usage when not in DB
In some cases when a scan process is in progress (or not started yet),
the root folder might have a size of -1.

To make sure that the quota can still be calculated correctly, an
special SQL query is used to pre-calculate the root folder size by
summing the size of all known files, which is better than no value at
all.

Fixes #5509